refactor: moving the TextareaAutosize from mui-material to mui-core
