+++ packages/mui-core/src/TextareaAutosize/TextareaAutosize.d.ts
