+++ packages/mui-core/src/TextareaAutosize/index.d.ts
