+++ packages/mui-core/src/TextareaAutosize/index.js
